Restructure and correct imports
