+++ src/membrain_seg/segmentation/dataloading/memseg_dataset.py
@@ -6,8 +6,8 @@ import imageio as io
 import numpy as np
 from torch.utils.data import Dataset
 
-from membrain_seg.dataloading.data_utils import read_nifti
-from membrain_seg.dataloading.memseg_augmentation import (
+from membrain_seg.segmentation.dataloading.data_utils import read_nifti
+from membrain_seg.segmentation.dataloading.memseg_augmentation import (
     get_training_transforms,
     get_validation_transforms,
 )
