+++ src/membrain_seg/segmentation/parse_utils.py
