+++ src/membrain_seg/segmentation/training/optim_utils.py
@@ -101,10 +101,10 @@ class DeepSuperVisionLoss(_Loss):
     Deep Supervision loss using downsampled GT and low-res outputs.
 
     Implementation based on nnU-Net's implementation with downsampled images.
-    Reference: Zeng, Guodong, et al. "3D U-net with multi-level deep supervision: 
-    fully automatic segmentation of proximal femur in 3D MR images." Machine Learning 
-    in Medical Imaging: 8th International Workshop, MLMI 2017, Held in Conjunction with 
-    MICCAI 2017, Quebec City, QC, Canada, September 10, 2017, Proceedings 8. Springer 
+    Reference: Zeng, Guodong, et al. "3D U-net with multi-level deep supervision:
+    fully automatic segmentation of proximal femur in 3D MR images." Machine Learning
+    in Medical Imaging: 8th International Workshop, MLMI 2017, Held in Conjunction with
+    MICCAI 2017, Quebec City, QC, Canada, September 10, 2017, Proceedings 8. Springer
     International Publishing, 2017.
 
     Parameters
