+++ src/membrain_seg/tomo_preprocessing/amplitude_spectrum_matching/_cli.py
@@ -1,5 +1,6 @@
-from membrain_seg.dataloading.data_utils import load_tomogram
 from typer import Option
+
+from membrain_seg.dataloading.data_utils import load_tomogram
 
 from ..cli import OPTION_PROMPT_KWARGS as PKWARGS
 from ..cli import cli
